Re-design SKILLS page
Fully redesigned SKILLS page, to match
desired skills according to my future position.

Right now it serves as a template for the main
skills branches, which I want to achive. I want to
have it like a path which I want to focus.

It will be updated along with investigation of
set of skills ineach branch.